Failed try to test license plate length for Task 5.
