+++ ClassLibrary/MC.cs
@@ -14,7 +14,7 @@ namespace ClassLibrary
         /// <summary>
         /// The License plate of the motorcycle
         /// </summary>
-        public string Licenseplate { get; set; }
+        public string LicensePlate { get; set; }
 
         /// <summary>
         /// The date and time the ticket is issued.
