+++ TestTicketingSystem/CarTest.cs
@@ -1,3 +1,4 @@
+using System;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using ClassLibrary;
 
@@ -7,7 +8,7 @@ namespace TestTicketingSystem
     public class CarTest
     {
         [TestMethod]
-        public void PriceTest()
+        public void CarPriceTest()
         {
             //Arrange
             Car car = new Car();
@@ -19,7 +20,7 @@ namespace TestTicketingSystem
         }
 
         [TestMethod]
-        public void VehicleTypeTest()
+        public void CarVehicleTypeTest()
         {
             //Arrange
             Car car = new Car();
@@ -29,5 +30,41 @@ namespace TestTicketingSystem
             //Assert
             Assert.AreEqual(expectedVehicleType, actualVehicleType);
         }
-    }
+
+        [TestMethod]
+        [ExpectedException(typeof(Exception),"License Plate must be maximum 7 characters.")]
+        public void MoreThan7LicensePlateTest()
+        {
+            //Arrange
+            Car car = new Car();
+            //Act
+            string LP = "12345678";
+            //Assert
+            Assert.AreEqual(LP, car.LicensePlate);
+        }
+
+        [TestMethod]
+        public void MCPriceTest()
+        {
+            //Arrange
+            MC mc = new MC();
+            //Act
+            var actualPrice = mc.Price();
+            double expectedPrice = 125;
+            //Assert
+            Assert.AreEqual(expectedPrice, actualPrice);
+        }
+
+        [TestMethod]
+        public void MCVehicleTypeTest()
+        {
+            //Arrange
+            MC mc = new MC();
+            //Act
+            var actualVehicleType = mc.VehicleType();
+            string expectedVehicleType = "MC";
+            //Assert
+            Assert.AreEqual(expectedVehicleType, actualVehicleType);
+        }
+        }
 }
